Move file format tests to separate test class.
